requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
textblob==0.17.1
pyttsx3==2.90
tensorflow==2.13.0
tensorflow-datasets==4.9.3
streamlit==1.28.1
pillow==10.0.1
scipy==1.11.4
matplotlib==3.7.2
scikit-learn==1.3.2
